# Improve speed linting