--- a/.eslintrc
+++ b/.eslintrc
@@ -1,41 +1,35 @@
 {
-  "extends": [
-    "standard-with-typescript",
-    "standard",
-    "standard-jsx",
-    "standard-react",
-    "next",
-    "prettier",
-    "prettier/react",
-    "prettier/@typescript-eslint"
-  ],
-  "plugins": ["simple-import-sort", "unicorn"],
-  "rules": {
-    "react/react-in-jsx-scope": "off",
-    "react/prop-types": "off",
-    "unicorn/no-abusive-eslint-disable": "error",
-    "no-console": "warn",
-    "react/display-name": "warn",
-    "react-hooks/rules-of-hooks": "warn",
-    "@typescript-eslint/triple-slash-reference": "off",
-    "@typescript-eslint/strict-boolean-expressions": "off",
-    "@typescript-eslint/explicit-function-return-type": "off",
-    "@typescript-eslint/consistent-type-definitions": "off",
-    "@typescript-eslint/restrict-template-expressions": "off",
-    "@typescript-eslint/no-floating-promises": "off",
-    "@typescript-eslint/return-await": "off",
-    "@typescript-eslint/consistent-type-imports": "error",
-    "simple-import-sort/imports": "error",
-    "simple-import-sort/exports": "error"
-  },
-  "globals": {
-    "fetch": false,
-    "localStorage": "readonly"
-  },
   "env": {
+    "browser": true,
+    "es2021": true,
+    "node": true,
     "jest": true
   },
+  "extends": [
+    "eslint:recommended",
+    "plugin:react/recommended",
+    "plugin:@typescript-eslint/recommended",
+    "next",
+    "prettier"
+  ],
+  "parser": "@typescript-eslint/parser",
   "parserOptions": {
-    "project": "./tsconfig.json"
+    "ecmaFeatures": {
+      "jsx": true
+    },
+    "ecmaVersion": 12,
+    "sourceType": "module"
+  },
+  "plugins": ["simple-import-sort", "react", "@typescript-eslint", "unicorn"],
+  "rules": {
+    "@typescript-eslint/consistent-type-imports": "error",
+    "simple-import-sort/imports": "error",
+    "simple-import-sort/exports": "error",
+    "no-console": "warn",
+    "react/react-in-jsx-scope": "off",
+    "react/prop-types": "off",
+    "@typescript-eslint/no-empty-function": "off",
+    "@typescript-eslint/explicit-module-boundary-types": "off",
+    "@typescript-eslint/ban-ts-comment": "off"
   }
 }