--- a/src/test/utils/setupPage.js
+++ b/src/test/utils/setupPage.js
@@ -1,11 +1,9 @@
-import router from 'next/router';
-
 import { match } from 'path-to-regexp';
 
 import { user } from '@/test/data/user';
 
 const mockRouterWithParams = ({ query, ...rest } = {}) => {
-  const useRouter = jest.spyOn(router);
+  const useRouter = jest.spyOn(require('next/router'), 'useRouter');
 
   const push = jest.fn();