--- a/src/ui/Box.tsx
+++ b/src/ui/Box.tsx
@@ -101,7 +101,10 @@ type TypeaheadProps = {
   isInvalid: boolean;
   allowNew:
     | boolean
-    | ((results: Array<Object | string>, props: Object) => boolean);
+    | ((
+        results: Array<Record<string, unknown> | string>,
+        props: Record<string, unknown>,
+      ) => boolean);
   newSelectionPrefix: string;
   inputProps: HTMLProps<HTMLInputElement>;
 };