Context:

My vscode environment has become really slow at times. It seems to be the eslint that runs on save.
I started this PR to improve the speed of our linting without losing out on features.

These are my findings using the `TIMING=1 yarn lint` command

**Before change:**

```
Rule                                      | Time (ms) | Relative
:-----------------------------------------|----------:|--------:
@typescript-eslint/no-misused-promises    |  3368.998 |    34.5%
@typescript-eslint/promise-function-async |  3061.135 |    31.3%
@typescript-eslint/no-implied-eval        |  1388.643 |    14.2%
import/no-duplicates                      |   527.297 |     5.4%
react/display-name                        |   106.398 |     1.1%
simple-import-sort/imports                |    91.702 |     0.9%
react/no-deprecated                       |    86.960 |     0.9%
@typescript-eslint/no-base-to-string      |    84.452 |     0.9%
react/no-direct-mutation-state            |    70.299 |     0.7%
no-unused-vars                            |    61.109 |     0.6%
✨  Done in 19.98s.
```

**After change:**

```
@typescript-eslint/no-unused-vars |   296.503 |    24.9%
react/display-name                |   121.027 |    10.1%
react/no-deprecated               |   117.768 |     9.9%
simple-import-sort/imports        |   111.195 |     9.3%
react/no-direct-mutation-state    |    92.569 |     7.8%
react/no-string-refs              |    57.236 |     4.8%
react/require-render-return       |    50.311 |     4.2%
react/no-unknown-property         |    30.897 |     2.6%
no-redeclare                      |    29.949 |     2.5%
react-hooks/exhaustive-deps       |    20.351 |     1.7%
✨  Done in 7.45s.
```