@geoffreydhuyvetters 2021-11-22
Could you explain in the description what the exact strategy is here? Or where this is based upon?