@simon-debruijn 2021-11-22
> Could you explain in the description what the exact strategy is here? Or where this is based upon?

I removed the eslint config we had, started from scratch with the `yarn eslint --init` command and started adding extends and plugins along the way.

The way typescript is checked, I think, has the biggest influence on the performance of our linting.
 
I could re-add the `standard` rules we had, from their website it seemed like they really want you to use standard instead of eslint so that's why I left them out for now.

I also did a `depcheck` and removed some unused dependencies.